Update and get claims endpoint
