+++ apps/web/src/utils/urls.ts
@@ -40,7 +40,6 @@ export const urls = {
       assetMint: '/api/v1/asset/mint',
       assetRedeem: '/api/v1/asset/redeem',
       assetTransfer: '/api/v1/asset/transfer',
-      claims: '/api/v1/admin/claims',
       createBankAccount: '/api/v1/payments/create-bank-account',
       createBidForPack: '/api/v1/bid/add-to-pack',
       createCard: '/api/v1/payments/create-card',
@@ -54,6 +53,7 @@ export const urls = {
       getBankAccountStatus: '/api/v1/payments/get-bank-account-status',
       getCardsByOwner: '/api/v1/payments/get-cards-by-owner',
       getCardStatus: '/api/v1/payments/get-card-status',
+      getClaims: '/api/v1/admin/get-claims',
       getPayment: '/api/v1/payments/get-payment',
       getPublishedPacks: '/api/v1/pack/get-published-packs',
       getRedeemable: '/api/v1/asset/get-redeemable',
@@ -64,6 +64,7 @@ export const urls = {
       removeCard: '/api/v1/payments/remove-card',
       showcaseCollectible: '/api/v1/collection/collectibles-showcase',
       updateCard: '/api/v1/payments/update-card',
+      updateClaims: '/api/v1/admin/update-claims',
       updateEmail: '/api/v1/profile/update-email',
       updateUsername: '/api/v1/profile/update-username',
       verifyPassphrase: '/api/v1/profile/verify-passphrase',
